# Improves ReduceSum performance by removing transposition.