… onnxruntime/core/providers/cpu/reduction/reduction_ops.h
… orttraining/orttraining/training_ops/cpu/reduction/reduction_ops.cc