**Description**

- Makes ReduceSum 4 times faster when the tensor is transposed, does not change when it is not transposed.
- Addresses issue #5253.
- The new implementation never transposes the tensor before reduction.

**Motivation and Context**

See issue #5253.

This changes only modifies ReduceSum implementation but it could be easily extended to other Reduce functions.

[test_reduce_sum.zip](https://github.com/microsoft/onnxruntime/files/5327345/test_reduce_sum.zip)

Before: 

```
case 0 shape=[5, 5, 2] axes=[0] ort=0.00059s tf 0.02617s size=50 ratio=0.02
case 1 shape=[5, 5, 2] axes=[1] ort=0.00029s tf 0.00316s size=50 ratio=0.09
case 2 shape=[5, 5, 2] axes=[2] ort=0.00161s tf 0.00106s size=50 ratio=1.52
case 3 shape=[5, 5, 2] axes=[0, 2] ort=0.00034s tf 0.00279s size=50 ratio=0.12
case 4 shape=[5, 5, 2] axes=[1, 2] ort=0.00029s tf 0.00076s size=50 ratio=0.38
case 5 shape=[5, 5, 2] axes=[0, 1, 2] ort=0.00025s tf 0.00141s size=50 ratio=0.18
case 6 shape=[5, 5, 2] axes=[0, 1] ort=0.00036s tf 0.00072s size=50 ratio=0.50
case 7 shape=[1000, 50, 20, 10] axes=[0] ort=0.17625s tf 0.23279s size=10000000 ratio=0.76
case 8 shape=[1000, 50, 20, 10] axes=[1] ort=0.19887s tf 0.06070s size=10000000 ratio=3.28
case 9 shape=[1000, 50, 20, 10] axes=[2] ort=0.42747s tf 0.06571s size=10000000 ratio=6.51
case 10 shape=[1000, 50, 20, 10] axes=[3] ort=0.06898s tf 0.04233s size=10000000 ratio=1.63
case 11 shape=[1000, 50, 20, 10] axes=[0, 3] ort=0.85743s tf 0.22630s size=10000000 ratio=3.79
case 12 shape=[1000, 50, 20, 10] axes=[1, 2] ort=0.29788s tf 0.04794s size=10000000 ratio=6.21
case 13 shape=[1000, 50, 20, 10] axes=[2, 3] ort=0.01832s tf 0.04093s size=10000000 ratio=0.45
case 14 shape=[1000, 50, 20, 10] axes=[1, 2, 3] ort=0.01812s tf 0.04167s size=10000000 ratio=0.43
case 15 shape=[1000, 50, 20, 10] axes=[0, 1, 2, 3] ort=0.02449s tf 0.02052s size=10000000 ratio=1.19
case 16 shape=[1000, 50, 20, 10] axes=[0, 1, 2] ort=0.20615s tf 0.19024s size=10000000 ratio=1.08
case 17 shape=[1000, 50, 20, 10] axes=[0, 1] ort=0.36285s tf 0.71741s size=10000000 ratio=0.51
case 18 shape=[100, 200, 4096] axes=[0] ort=1.60119s tf 0.74998s size=81920000 ratio=2.13
case 19 shape=[100, 200, 4096] axes=[1] ort=1.64074s tf 0.83068s size=81920000 ratio=1.98
case 20 shape=[100, 200, 4096] axes=[2] ort=0.13331s tf 0.34629s size=81920000 ratio=0.38
case 21 shape=[100, 200, 4096] axes=[0, 2] ort=8.33892s tf 0.78521s size=81920000 ratio=10.62
case 22 shape=[100, 200, 4096] axes=[1, 2] ort=0.14292s tf 0.33101s size=81920000 ratio=0.43
case 23 shape=[100, 200, 4096] axes=[0, 1, 2] ort=0.18970s tf 0.13298s size=81920000 ratio=1.43
case 24 shape=[100, 200, 4096] axes=[0, 1] ort=2.87974s tf 4.72749s size=81920000 ratio=0.61
case 25 shape=[10, 10, 512, 512] axes=[0] ort=0.41671s tf 0.13075s size=26214400 ratio=3.19
case 26 shape=[10, 10, 512, 512] axes=[1] ort=0.41590s tf 0.23847s size=26214400 ratio=1.74
case 27 shape=[10, 10, 512, 512] axes=[2] ort=0.55169s tf 0.33318s size=26214400 ratio=1.66
case 28 shape=[10, 10, 512, 512] axes=[3] ort=0.04332s tf 0.09974s size=26214400 ratio=0.43
case 29 shape=[10, 10, 512, 512] axes=[1, 2] ort=0.91328s tf 1.56667s size=26214400 ratio=0.58
case 30 shape=[10, 10, 512, 512] axes=[0, 3] ort=7.39572s tf 0.11926s size=26214400 ratio=62.01
case 31 shape=[10, 10, 512, 512] axes=[2, 3] ort=0.04093s tf 0.10728s size=26214400 ratio=0.38
case 32 shape=[10, 10, 512, 512] axes=[1, 2, 3] ort=0.04104s tf 0.29138s size=26214400 ratio=0.14
case 33 shape=[10, 10, 512, 512] axes=[0, 1, 2, 3] ort=0.07740s tf 0.04589s size=26214400 ratio=1.69
case 34 shape=[10, 10, 512, 512] axes=[0, 1, 2] ort=0.98779s tf 1.64959s size=26214400 ratio=0.60
case 35 shape=[10, 10, 512, 512] axes=[0, 1] ort=0.49940s tf 0.24972s size=26214400 ratio=2.00
case 36 shape=[2, 2, 256, 256, 256] axes=[0] ort=1.71587s tf 0.55286s size=67108864 ratio=3.10
case 37 shape=[2, 2, 256, 256, 256] axes=[1] ort=1.74283s tf 1.81400s size=67108864 ratio=0.96
case 38 shape=[2, 2, 256, 256, 256] axes=[2] ort=1.36341s tf 0.68762s size=67108864 ratio=1.98
case 39 shape=[2, 2, 256, 256, 256] axes=[3] ort=1.57968s tf 0.45804s size=67108864 ratio=3.45
case 40 shape=[2, 2, 256, 256, 256] axes=[4] ort=0.12364s tf 0.25359s size=67108864 ratio=0.49
case 41 shape=[2, 2, 256, 256, 256] axes=[1, 3] ort=1.60568s tf 2.25171s size=67108864 ratio=0.71
case 42 shape=[2, 2, 256, 256, 256] axes=[0, 4] ort=18.98522s tf 0.32439s size=67108864 ratio=58.53
case 43 shape=[2, 2, 256, 256, 256] axes=[3, 4] ort=0.13100s tf 0.24954s size=67108864 ratio=0.52
case 44 shape=[2, 2, 256, 256, 256] axes=[2, 3, 4] ort=0.11128s tf 0.93328s size=67108864 ratio=0.12
case 45 shape=[2, 2, 256, 256, 256] axes=[1, 2, 3, 4] ort=0.10885s tf 0.95819s size=67108864 ratio=0.11
case 46 shape=[2, 2, 256, 256, 256] axes=[0, 1, 2, 3, 4] ort=0.16012s tf 0.10923s size=67108864 ratio=1.47
case 47 shape=[2, 2, 256, 256, 256] axes=[0, 1, 2, 3] ort=2.46403s tf 4.51959s size=67108864 ratio=0.55
case 48 shape=[2, 2, 256, 256, 256] axes=[0, 1, 2] ort=1.85099s tf 4.27360s size=67108864 ratio=0.43
case 49 shape=[2, 2, 256, 256, 256] axes=[0, 1] ort=1.30402s tf 0.37666s size=67108864 ratio=3.46
```

After:

```
case 0 shape=[5, 5, 2] axes=[0] ort=0.00061s tf 0.01442s size=50 ratio=0.04 speedup=0.50
case 1 shape=[5, 5, 2] axes=[1] ort=0.00043s tf 0.00284s size=50 ratio=0.15 speedup=0.60
case 2 shape=[5, 5, 2] axes=[2] ort=0.00034s tf 0.00075s size=50 ratio=0.46 speedup=3.30
case 3 shape=[5, 5, 2] axes=[0, 2] ort=0.00029s tf 0.00190s size=50 ratio=0.15 speedup=0.80
case 4 shape=[5, 5, 2] axes=[1, 2] ort=0.00028s tf 0.00082s size=50 ratio=0.34 speedup=1.12
case 5 shape=[5, 5, 2] axes=[0, 1, 2] ort=0.00032s tf 0.00167s size=50 ratio=0.19 speedup=0.95
case 6 shape=[5, 5, 2] axes=[0, 1] ort=0.00027s tf 0.00102s size=50 ratio=0.27 speedup=1.85
case 7 shape=[1000, 50, 20, 10] axes=[0] ort=0.10830s tf 0.23524s size=10000000 ratio=0.46 speedup=1.65
case 8 shape=[1000, 50, 20, 10] axes=[1] ort=0.07417s tf 0.05973s size=10000000 ratio=1.24 speedup=2.65
case 9 shape=[1000, 50, 20, 10] axes=[2] ort=0.05709s tf 0.06646s size=10000000 ratio=0.86 speedup=7.57
case 10 shape=[1000, 50, 20, 10] axes=[3] ort=0.06412s tf 0.04068s size=10000000 ratio=1.58 speedup=1.03
case 11 shape=[1000, 50, 20, 10] axes=[0, 3] ort=0.09211s tf 0.22908s size=10000000 ratio=0.40 speedup=9.47
case 12 shape=[1000, 50, 20, 10] axes=[1, 2] ort=0.08035s tf 0.04648s size=10000000 ratio=1.73 speedup=3.59
case 13 shape=[1000, 50, 20, 10] axes=[2, 3] ort=0.01881s tf 0.04074s size=10000000 ratio=0.46 speedup=0.98
case 14 shape=[1000, 50, 20, 10] axes=[1, 2, 3] ort=0.01767s tf 0.03991s size=10000000 ratio=0.44 speedup=0.98
case 15 shape=[1000, 50, 20, 10] axes=[0, 1, 2, 3] ort=0.02337s tf 0.02016s size=10000000 ratio=1.16 speedup=1.03
case 16 shape=[1000, 50, 20, 10] axes=[0, 1, 2] ort=0.23828s tf 0.20885s size=10000000 ratio=1.14 speedup=0.95
case 17 shape=[1000, 50, 20, 10] axes=[0, 1] ort=0.87929s tf 0.71663s size=10000000 ratio=1.23 speedup=0.41
case 18 shape=[100, 200, 4096] axes=[0] ort=0.98415s tf 0.73656s size=81920000 ratio=1.34 speedup=1.59
case 19 shape=[100, 200, 4096] axes=[1] ort=1.02354s tf 0.81880s size=81920000 ratio=1.25 speedup=1.58
case 20 shape=[100, 200, 4096] axes=[2] ort=0.13566s tf 0.30922s size=81920000 ratio=0.44 speedup=0.86
case 21 shape=[100, 200, 4096] axes=[0, 2] ort=1.26296s tf 0.88744s size=81920000 ratio=1.42 speedup=7.48
case 22 shape=[100, 200, 4096] axes=[1, 2] ort=0.13375s tf 0.34534s size=81920000 ratio=0.39 speedup=1.10
case 23 shape=[100, 200, 4096] axes=[0, 1, 2] ort=0.20135s tf 0.13306s size=81920000 ratio=1.51 speedup=0.95
case 24 shape=[100, 200, 4096] axes=[0, 1] ort=7.89768s tf 4.77809s size=81920000 ratio=1.65 speedup=0.37
case 25 shape=[10, 10, 512, 512] axes=[0] ort=0.22716s tf 0.13233s size=26214400 ratio=1.72 speedup=1.85
case 26 shape=[10, 10, 512, 512] axes=[1] ort=0.23386s tf 0.24030s size=26214400 ratio=0.97 speedup=1.79
case 27 shape=[10, 10, 512, 512] axes=[2] ort=0.37608s tf 0.33957s size=26214400 ratio=1.11 speedup=1.50
case 28 shape=[10, 10, 512, 512] axes=[3] ort=0.04437s tf 0.10144s size=26214400 ratio=0.44 speedup=0.98
case 29 shape=[10, 10, 512, 512] axes=[1, 2] ort=1.67736s tf 1.48470s size=26214400 ratio=1.13 speedup=0.51
case 30 shape=[10, 10, 512, 512] axes=[0, 3] ort=0.21570s tf 0.11688s size=26214400 ratio=1.85 speedup=33.52
case 31 shape=[10, 10, 512, 512] axes=[2, 3] ort=0.04180s tf 0.10829s size=26214400 ratio=0.39 speedup=0.97
case 32 shape=[10, 10, 512, 512] axes=[1, 2, 3] ort=0.04129s tf 0.29397s size=26214400 ratio=0.14 speedup=1.00
case 33 shape=[10, 10, 512, 512] axes=[0, 1, 2, 3] ort=0.06181s tf 0.04351s size=26214400 ratio=1.42 speedup=1.19
case 34 shape=[10, 10, 512, 512] axes=[0, 1, 2] ort=2.78419s tf 1.69213s size=26214400 ratio=1.65 speedup=0.36
case 35 shape=[10, 10, 512, 512] axes=[0, 1] ort=0.31127s tf 0.24751s size=26214400 ratio=1.26 speedup=1.59
case 36 shape=[2, 2, 256, 256, 256] axes=[0] ort=1.56215s tf 0.55650s size=67108864 ratio=2.81 speedup=1.10
case 37 shape=[2, 2, 256, 256, 256] axes=[1] ort=1.53227s tf 1.80231s size=67108864 ratio=0.85 speedup=1.13
case 38 shape=[2, 2, 256, 256, 256] axes=[2] ort=0.89286s tf 0.69976s size=67108864 ratio=1.28 speedup=1.55
case 39 shape=[2, 2, 256, 256, 256] axes=[3] ort=0.59611s tf 0.48241s size=67108864 ratio=1.24 speedup=2.78
case 40 shape=[2, 2, 256, 256, 256] axes=[4] ort=0.12978s tf 0.25315s size=67108864 ratio=0.51 speedup=0.96
case 41 shape=[2, 2, 256, 256, 256] axes=[1, 3] ort=0.86158s tf 2.27005s size=67108864 ratio=0.38 speedup=1.87
case 42 shape=[2, 2, 256, 256, 256] axes=[0, 4] ort=0.44899s tf 0.31524s size=67108864 ratio=1.42 speedup=41.22
case 43 shape=[2, 2, 256, 256, 256] axes=[3, 4] ort=0.11124s tf 0.25101s size=67108864 ratio=0.44 speedup=1.18
case 44 shape=[2, 2, 256, 256, 256] axes=[2, 3, 4] ort=0.12756s tf 0.92857s size=67108864 ratio=0.14 speedup=0.86
case 45 shape=[2, 2, 256, 256, 256] axes=[1, 2, 3, 4] ort=0.10977s tf 0.92960s size=67108864 ratio=0.12 speedup=0.92
case 46 shape=[2, 2, 256, 256, 256] axes=[0, 1, 2, 3, 4] ort=0.16158s tf 0.11016s size=67108864 ratio=1.47 speedup=1.00
case 47 shape=[2, 2, 256, 256, 256] axes=[0, 1, 2, 3] ort=6.70774s tf 4.46010s size=67108864 ratio=1.50 speedup=0.37
case 48 shape=[2, 2, 256, 256, 256] axes=[0, 1, 2] ort=2.49528s tf 4.24235s size=67108864 ratio=0.59 speedup=0.73
case 49 shape=[2, 2, 256, 256, 256] axes=[0, 1] ort=0.88900s tf 0.39538s size=67108864 ratio=2.25 speedup=1.54
```